xml to text

+++ xml_to_text.py
@@ -7,7 +7,7 @@ Created on Thu Feb 11 17:16:17 2021
 """
 
 import xml.etree.ElementTree as ET
-import pdb
+
 keep = [
         "title-group",
         "article-title",
@@ -22,26 +22,10 @@ keep = [
         "italic"
         ]
 
-#ignore = ["ext-link", "xref"]
 
-def print_text_ignore(el, file):
-    if el.tag in ignore:
-        return
-    if type(el.text) is str:
-        with open (file, 'a') as f:
-            f.write(el.text + "\n")
-    for child in el:
-        print_text_ignore(child, file)
-
-def handle_abstract(el):
-    for elc in el.iterfind(".*"):
-       print(elc.text)
-    pdb.set_trace()
 
 def print_text_keep(el, file):
-    #if el.tag == "abstract":
-     #  handle_abstract(el)
-    
+  
     if (el.tag in keep) and (type(el.text) is str):
         with open (file, 'a') as f:
             f.write(el.text + "\n")
@@ -51,18 +35,31 @@ def print_text_keep(el, file):
         if type(child.tail) is str and (el.tag in keep):
             with open(file, 'a') as f:
                 f.write(child.tail + "\n")
+                
+#file = 'converted_new.txt'
+#with open(file, 'w') as f:
+#    f.write("")
 
-file = '/Users/shalinimustala/Documents/Journal_analysis/allpapersTXT/10.1177_0001839214523602_3.txt' #'converted_new.txt'
-with open(file, 'w') as f:
-    f.write("")
-
-tree = ET.parse('/Users/shalinimustala/Documents/Journal_analysis/all_papers/10.1177_0001839214523602.xml')    #ET parser helps in iterating the tree
-root = tree.getroot()
-for child in root:
-   if child.tag != 'back':
-        print_text_keep(child, file)
+#tree = ET.parse('paper.xml')    #ET parser helps in iterating the tree
+#root = tree.getroot()
+#for child in root:
+#    if child.tag != 'back':
+#        print_text_keep(child, file)
         
-
+def allpapers(in_dir, out_dir):
+    import os
+    #directory = os.getcwd() + '/all_papers/'
+    for entry in os.scandir(in_dir):
+        if entry.path.endswith(".xml") and entry.is_file():
+            path_paper = entry.path
+            file = out_dir + "/" + path_paper.split("/")[-1].replace("xml", "txt")
+            with open(file, 'w') as f:
+                f.write("")
+            tree = ET.parse(path_paper)
+            root = tree.getroot()
+            for child in root:
+                if child.tag != 'back':
+                    print_text_keep(child, file)
